Uncomment commented parameters in inlined functions
The code had a lot of cases where parameters of inlined functions were set in
comments. This way, the doxy documentation couldn't pick up the variables
correctly (because the identification is based on parameter names). So I
uncommented all commented parameters. Not sure why this was done in the first
place ...