Merge pre-build of BAT and build steps into one and make them manual
We don't actually need to test all three ways to build KLFitter for _every_
pipeline. The standard way should be to include BAT in the build process and use
cmake to make KLFitter. The other options are now simplified (no pre-build step
anymore) and they can be triggered manually if needed.