Re-adding delete commands for MET resolution object
We will need to think about a nicer way of dealing with these resolution
pointers, because in the case of MET we don't have eta parameterisations, and
therefore no smart pointers in the derived classes. For now, the delete command
for the MET resolution has been re-added to the detector classes.