Declare core classes as 'final'
The core classes (Fitter, Particles, Permutations, PhysicsConstants) do not have
virtual functions implemented, and there is no intention to ever derive classes
from them. To avoid this, they are now marked as 'final' which will lead to a
compilation error if one of them is used as a base class. Consequently, the
'virtual' keyword was removed from the destructors.