Remove 'virtual' keyword where not needed
Many methods of the likelihood classes carry the 'virtual' keyword, although
they are never overwritten (also because almost none of them are used as parent
classes). The 'virtual' keyword was removed from all methods where it was not
necessary. With one exception, this only leaves virtual methods in the three
"Base" classes: LikelihoodBase, ResolutionBase and DetectorBase.