Declare LikelihoodBase functions purely virtual if appropriate
A lot of the member functions are re-implemented in the derived classes (and
also only make sense if re-implemented). Instead of returning dummy values,
nullpointers etc. the functions are now declared purely virtual in the base
class. That is, if not reimplemented in a derived class, the compilation will
fail.