Remove checks for permutations.empty() is those cases are expected
With the change from pointers to objects, the checks if (!permutationTable) were
changed into if (permutationTable.empty()), which is not the identical check. We
expect cases where the table is actually empty, e.g. if all permutations are
vetoed, but we don't want those cases to issue confusing warnings. The
std::vector<T>::empty() checks in Permutations.cxx were therefore removed.