Move from std::unique_ptr<TLorentzVector> to objects
This change was applied especially to the containers used in the Particles
class, i.e. of type std::vector<std::unique_ptr<TLorentzVector> >. It makes
handling of the objects a lot easier in many cases. The only drawback is that we
need to include TLorentzVector.h in the header file now.