Remove constexpr specifier from expressions with std library
For some reason, these expressions never caused problems on machines that we
built KLFitter on. But when using clang on mac, the compiler gives errors for
these expressions as functions, such as std::sqrt(), are not const expressinos
themselves. To follow c++ standards, we now converted all of these expressions
into `const` only.